Remove serverless as deployment method for Lamdbas:
It adds to much overhead and we cannot reference the ZIP files it
uploads in other stacks, as we don't know the S3 path to them.

But we need need the path to create deployment functions per deployment
pipeline, so that it works with multiple pipelines derived from
`pipeline-serverless-backend-npm-frontend` and also in multi account
setups.

 - use a node lib which provides a zip cli as there is no local ZIP
   binary
 - use npm package to do the s3 deployment of Lambda zip package, in
   order to keep size of source code below 4096